# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: URJOB API
  description: "URJOB 프로젝트의 전체 API 명세 (Backend + BFF)"
  version: "1.2.0"
servers:
  - url: http://localhost:8081  # BFF 호스트 (dev)
    description: BFF (dev)
  - url: http://localhost:8082  # Backend 호스트 (dev)
    description: Backend (dev)
...
paths:
...
  /introductions/{id}:
    get:
      summary: 자기소개서 상세 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 자기소개서 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
    put:
      summary: 자기소개서 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntroductionCreate'
      responses:
        "200":
          description: 자기소개서 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
    delete:
      summary: 자기소개서 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 자기소개서 삭제 성공

  # ---------------------------
  # Community (사용자 커뮤니티)
  # ---------------------------
...
  /community/posts/{id}:
    get:
      summary: 게시글 상세 조회 (조회수 증가)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 게시글 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostDetail'
    put:
      summary: 게시글 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityPostCreate'
      responses:
        "200":
          description: 게시글 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostDetail'
    delete:
      summary: 게시글 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 게시글 삭제 성공
...
components:
  schemas:
...
    Introduction:
      type: object
      properties:
        id:
          type: number
          description: 자기소개서의 고유 ID
        jobSeekerId:
          type: number
          description: 구직자의 ID
        title:
          type: string
          description: 자기소개서 제목
        content:
          type: string
          description: 자기소개서 내용
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - jobSeekerId
        - title
        - content
        - createdAt
        - updatedAt
    IntroductionCreate:
      type: object
      properties:
        jobSeekerId:
          type: number
        title:
          type: string
        content:
          type: string
      required:
        - jobSeekerId
        - title
        - content
...
    CommunityPostDetail:
      type: object
      properties:
        id:
          type: number
          description: community_posts.post_id
        subTitle:
          type: string
          description: "게시글 내용 요약 또는 부제목"
        title:
          type: string
        writer:
          type: string
        date:
          type: string
          format: date-time
        views:
          type: integer
        likes:
          type: integer
        content:
          type: string
          description: community_posts.content
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommunityComment'
      required:
        - id
        - title
        - writer
        - date
        - views
        - likes
        - content
        - comments
    CommunityPostCreate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        area:
          type: string
      required:
        - title
        - content
        - area
    CommunityComment:
      type: object
      properties:
        id:
          type: number
          description: community_comments.comment_id
        writer:
          type: string
          description: "댓글 작성자 이름"
        date:
          type: string
          format: date-time
          description: community_comments.created_at
        content:
          type: string
          description: community_comments.content
        likes:
          type: integer
          description: community_comments.likes
        recommentId:
          type: number
          description: "부모 댓글 ID (대댓글인 경우)"
      required:
        - id
        - writer
        - date
        - content
        - likes
...